Only allow single submissions
Repurposes the submissions table to only allow for a single message
submission to be approved. This is a breaking change, and will require
the original submissions table in the database to be dropped.

This also removes the AwardPointsModal because it was exactly the same
as Approve and people didn't even know it existed.